kitchensink: Add 'unused' symbols
The idea is to have symbols that we derive Facet for, but that we don't
use the shape of, because we're not serializing/deserializing anything
using them.

This ends up showing that removing `#[used]` in https://github.com/facet-rs/facet/pull/788
does result in savings in terms of symbol sizes (and LLVM lines).